requirements.txt
requests
selenium
webdriver-manager
SpeechRecognition 
pyaudio 
gTTS 
pygame 
langchain_groq 
langdetect

fastapi 
uvicorn 
python-multipart
pydub


python-dotenv
pydantic

coco-lib 
openai-whisper